+++ Tests/RudderStackAnalyticsTests/Base/AnalyticsTests.swift
@@ -12,7 +12,6 @@ import Testing
 // MARK: - Analytics Unit Tests
 
 @Suite("Analytics Unit Tests")
-@MainActor
 class AnalyticsTests {
     
     var analytics: Analytics
@@ -24,9 +23,9 @@ class AnalyticsTests {
         mockStorage = MockStorage()
         
         // Create analytics with mock configuration using our mock storage
-        let config = SwiftTestMockProvider.createMockConfiguration(
-            writeKey: SwiftTestMockProvider.mockWriteKey,
-            dataPlaneUrl: SwiftTestMockProvider.mockDataPlaneUrl,
+        let config = MockProvider.createMockConfiguration(
+            writeKey: MockProvider.mockWriteKey,
+            dataPlaneUrl: MockProvider.mockDataPlaneUrl,
             storage: mockStorage
         )
         
@@ -368,16 +367,12 @@ class AnalyticsTests {
     @Test("when calling flush, then flush is propagated to plugins")
     func testFlush() async {
         // Track some events first
-        analytics.track(name: "Test Event")
+        let mockAnalytics = MockAnalytics()
+        mockAnalytics.track(name: "Test Event")
         
-        await runAfter(0.1) {
-            #expect(self.mockStorage.batchCount == 1)
-            
-            self.analytics.flush()
-            await runAfter(0.3) {
-                #expect(self.mockStorage.batchCount == 0)
-            }
-        }
+        mockAnalytics.flush()
+        
+        #expect(mockAnalytics.isFlushed)
     }
     
     // MARK: - Deep Link Tests
