+++ Tests/RudderStackAnalyticsTests/EventsModule/EventProcessing/EventWriterTests.swift
@@ -10,7 +10,6 @@ import Testing
 @testable import RudderStackAnalytics
 
 @Suite("EventWriter Tests")
-@MainActor
 class EventWriterTests {
     var analytics: Analytics
     var mockStorage: MockStorage
@@ -20,7 +19,7 @@ class EventWriterTests {
     
     init() {
         mockStorage = MockStorage()
-        let config = SwiftTestMockProvider.createMockConfiguration(storage: mockStorage)
+        let config = MockProvider.createMockConfiguration(storage: mockStorage)
         config.trackApplicationLifecycleEvents = false
         config.sessionConfiguration.automaticSessionTracking = false
         config.flushPolicies = []
