+++ Tests/RudderStackAnalyticsTests/Network/HttpNetworkTests.swift
@@ -9,11 +9,12 @@ import Foundation
 import Testing
 @testable import RudderStackAnalytics
 
+#if !os(watchOS) // URLProtocol-based mocks don’t work on watchOS..
 @Suite("HttpNetwork Tests")
 class HttpNetworkTests {
     
-    init() { SwiftTestMockProvider.setupMockURLSession() }
-    deinit { SwiftTestMockProvider.teardownMockURLSession() }
+    init() { MockProvider.setupMockURLSession() }
+    deinit { MockProvider.teardownMockURLSession() }
     
     @Test("given a success request, when it returns success status code in range, then HttpNetwork returns expected data",
           arguments: [(200, "https://success.test.com"),
@@ -119,3 +120,4 @@ extension HttpNetworkTests {
         }
     }
 }
+#endif
